requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deepdiff==8.5.0
dm-env==1.6
dm-env-rpc==1.1.6
dm-tree==0.1.8
docopt-ng==0.9.0
enum34==1.1.10
estraces==1.9.6
eval_type_backport==0.3.1
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
grpcio==1.74.0
gym==0.25.1
gym-notices==0.1.0
gymnasium==1.1.1
h5py==3.14.0
idna==3.10
ImageIO==2.37.2
immutabledict==4.2.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.1
jsonpickle==4.1.1
kiwisolver==1.4.7
llvmlite==0.43.0
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.9.4
mock==5.2.0
mpmath==1.3.0
mpyq==0.2.5
munch==4.0.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
nvitop==1.5.2
orderly-set==5.5.0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
platformdirs==4.4.0
portpicker==1.6.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pyparsing==3.2.3
PySC2==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s2clientprotocol==5.0.14.93333.0
s2protocol==5.0.14.93333.0
sacred==0.8.7
scared==1.1.8
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.49.0
setproctitle==1.3.7
six==1.17.0
sk-video==1.1.10
# Editable install with no version control (SMAC==1.0.0)
-e /data/home/lianghj/SMAC_V1/smac
# Editable install with no version control (SMACv2==1.0.0)
-e /data/home/lianghj/SMAC_V2/smacv2
smmap==5.0.2
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
threadpoolctl==3.6.0
torch==2.4.1+cu121
torch_scatter==2.1.2+pt24cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
triton==3.0.0
trsfile==2.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.24.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
